DTSCCI-1585 changes
