+++ src/main/views/features/public/firstContact/claim-summary.njk
@@ -48,7 +48,7 @@
           </p>
           {{ govukDetails({
             summaryText: t('PAGES.FIRST_CONTACT_CLAIM_SUMMARY.VIEW_AMOUNT_BREAKDOWN'),
-            html: amountBreakdownTable(claim, interestData, totalAmount, t, fixedCost)
+            html: amountBreakdownTable(t, claim, interestData, totalAmount, fixedCost)
           }) }}
           <h2 class="govuk-heading-s">{{ t('PAGES.FIRST_CONTACT_CLAIM_SUMMARY.REASON_FOR_CLAIM') }}</h2>
           <p class="govuk-body">{{ claim.claimDetails.reason.text }}</p>
