+++ src/main/views/features/response/claimDetails/claim-details-new.njk
@@ -17,7 +17,7 @@
     <br>
     {{ govukDetails({
             summaryText: t('PAGES.CLAIM_DETAILS.VIEW_AMOUNT_BREAKDOWN'),
-            html: amountBreakdownTable(claim, interestData, totalAmount, t)
+            html: amountBreakdownTable(t, claim, interestData, totalAmount, fixedCost)
      }) }}
      <h2 class="govuk-heading-s">{{ t('PAGES.CLAIM_DETAILS.REASON_FOR_CLAIM')+' :' }}</h2>
      <p class="govuk-body">{{ claim.claimDetails.reason.text }}</p>
