+++ src/main/views/macro/amount-breakdown.njk
@@ -1,7 +1,7 @@
 {% from "govuk/components/details/macro.njk" import govukDetails %}
 {% from "./interest-snippet.njk" import interestSnippet %}
 
-{% macro amountBreakdownTable(claim, interestData, totalAmount, t) %}
+{% macro amountBreakdownTable(t, claim, interestData, totalAmount, fixedCost) %}
 <table class="govuk-table">
   <thead class="govuk-table__head">
     <tr class="govuk-table__row">
@@ -35,7 +35,7 @@
    {% if fixedCost !== undefined %}
      <tr class="govuk-table__row">
         <td class="govuk-table__cell">{{ t('PAGES.CLAIM_DETAILS.CLAIM_FIXED_COSTS') }}</td>
-        <td class="govuk-table__cell govuk-!-text-align-right">{{ fixedCost | pennies2pounds | currencyFormat(fixedCost) }}</td>
+        <td class="govuk-table__cell govuk-!-text-align-right">{{ fixedCost | currencyFormat(fixedCost) }}</td>
      </tr>
    {% endif %}
     <tr class="govuk-table__row">
